separates Applications concerns into audience, account and product and extract common logic into a module
